[Core] fix tests

+++ src/core/contracts/libraries/AdaptiveFee.sol
@@ -16,7 +16,6 @@ library AdaptiveFee {
         uint16 baseFee;
     }
 
-    // Maybe uint112
     function getFee(
         uint112 volatility,
         uint256 volumePerLiquidity,
@@ -25,7 +24,9 @@ library AdaptiveFee {
         uint256 sigm1 = sigmoid(volatility, config.gamma1, config.alpha1, config.beta1);
         uint256 sigm2 = sigmoid(volatility, config.gamma2, config.alpha2, config.beta2);
 
-        fee = config.baseFee + sigmoid(volumePerLiquidity, config.volumeGamma, sigm1 + sigm2, config.volumeBeta);
+        fee =
+            uint256(config.baseFee) +
+            sigmoid(volumePerLiquidity, config.volumeGamma, sigm1 + sigm2, config.volumeBeta);
     }
 
     function sigmoid(
